read_mesh function now able to read in triangles coords
read_mesh function in mesh.c will now read in triangle coordinates from mesh.input, convert them to a double and can print them out for each specific position. The next step is to save these coordinates to the array of triangles called mesh in order to be used for collision detection.