Reworked program output handling by removing data_out entirely
This changes makes it now impossible to time just the collision detection, as the program now outputs to file within the collide_all function. However, there is no longer any worry about the possiblity of a segmentation fault caused by the allocation of memory to pairs array, as we have bypassed this process entirely. This makes the program much more robust and prevents it from being limited in collision detection